+++ types/index.d.ts
@@ -19,6 +19,7 @@ import {
     YearNamingRules
 } from "../src/constants";
 import NoteStub from "../src/classes/notes/note-stub";
+import {currentDateTime, getCurrentDateTime} from "../src/classes/api";
 
 declare global{
     /**
@@ -206,6 +207,57 @@ declare global{
             export async function configureCalendar(calendarData: PredefinedCalendars | SimpleCalendar.CalendarData, calendarId: string = 'active'): Promise<boolean>
 
             /**
+             * Gets the current date and time for the current calendar or the passed in calendar.
+             * @param calendarId Optional parameter to specify the ID of the calendar to get the current day from. If not provided the current active calendar will be used.
+             *
+             * @returns The current date and time. The month and day are index's and as such start at 0 instead of 1.  If the passed in calendar can't be found null will be returned.
+             *
+             * @example
+             * ```javascript
+             * // Assuming a Gregorian calendar
+             * SimpleCalendar.api.currentDateTime();
+             * //Returns a DateTime object like this
+             * // {
+             * //     year: 2021,
+             * //     month: 11,
+             * //     day: 24,
+             * //     hour: 12,
+             * //     minute: 13,
+             * //     seconds: 14
+             * // }
+             * ```
+             */
+            export function currentDateTime(calendarId: string = 'active'): SimpleCalendar.DateTime | null
+
+            /**
+             * Gets the formatted display data for the current date and time of the active calendar, or the calendar with the passed in ID.
+             * @param calendarId Optional parameter to specify the ID of the calendar to get the current day from. If not provided the current active calendar will be used.
+             *
+             * @returns All the formatted display strings for the current date and time. Or if the passed in calendar can't be found, null.
+             *
+             * @example
+             * ```javascript
+             * // Assuming a Gregorian calendar
+             * SimpleCalendar.api.currentDateTimeDisplay();
+             * //Returns a DateTime object like this
+             * // {
+             * //     date: "June 01, 2021",
+             * //     day: "1",
+             * //     daySuffix: "st",
+             * //     month: "6",
+             * //     monthName: "June",
+             * //     time: "00:00:00",
+             * //     weekday: "Tuesday",
+             * //     year: "2021",
+             * //     yearName: "",
+             * //     yearPostfix: "",
+             * //     yearPrefix: ""
+             * // }
+             * ```
+             */
+            export function currentDateTimeDisplay(calendarId: string = 'active'): SimpleCalendar.DateDisplayData | null
+
+            /**
              * Converts the passed in date to a timestamp.
              * @param date A date object (eg `{year:2021, month: 4, day: 12, hour: 0, minute: 0, seconds: 0}`) with the parameters set to the date that should be converted to a timestamp. Any missing parameters will default to the current date value for that parameter.<br>**Important**: The month and day are index based so January would be 0 and the first day of the month will also be 0.
              * @param calendarId Optional parameter to specify the ID of the calendar to use when converting a date to a timestamp. If not provided the current active calendar will be used.
@@ -254,6 +306,30 @@ declare global{
              * ```
              */
             export function formatDateTime(date: SimpleCalendar.DateTimeParts, format: string = '', calendarId: string = 'active'): string | {date: string, time: string}
+
+            /**
+             * Converts the passed in timestamp into formatted date and time strings that match the configured date and time formats or the passed in format string.
+             *
+             * @param timestamp The timestamp (in seconds) of the date to format.
+             * @param format Optional format string to return custom formats for the passed in date and time.
+             * @param calendarId Optional parameter to specify the ID of the calendar to use when converting a date to a formatted string. If not provided the current active calendar will be used.
+             *
+             * @returns If no format string is provided an object with the date and time formatted strings, as set in the configuration, will be returned. If a format is provided then a formatted string will be returned.
+             *
+             * @examples
+             * ```javascript
+             * // Assuming that the default date and time formats are in place
+             * // Date: Full Month Name Day, Year
+             * // Time: 24Hour:Minute:Second
+             *
+             * SimpleCalendar.api.formatTimestamp(1640434394);
+             * // Returns {date: 'December 25, 2021', time: '12:13:14'}
+             *
+             * SimpleCalendar.api.formatTimestamp(1640434394,"DD/MM/YYYY HH:mm:ss A");
+             * // Returns '25/12/2021 12:13:14 PM'
+             * ```
+             */
+            export function formatTimestamp(timestamp: number, format: string = '', calendarId: string = 'active'): string | {date: string, time: string}
 
             /**
              * Gets the details of all calendars that have been configured in Simple Calendar